split off sign-only variant of CryptoSettingsDialog
