+++ k9mail/src/main/java/com/fsck/k9/activity/compose/RecipientPresenter.java
@@ -545,7 +545,7 @@ public class RecipientPresenter implements PermissionPingCallback {
                 Log.e(K9.LOG_TAG, "click on crypto status while unconfigured - this should not really happen?!");
                 return;
             case OK:
-                recipientMvpView.showCryptoDialog(currentCryptoMode);
+                recipientMvpView.showCryptoDialog(currentCryptoMode, account.getCryptoSupportSignOnly());
                 return;
 
             case LOST_CONNECTION:
