+++ k9mail/src/main/res/values/strings.xml
@@ -1155,10 +1155,13 @@ Please submit bug reports, contribute new features and ask questions at
     <string name="compose_error_pgp_error">Error with OpenPGP provider. Check your settings!</string>
     <string name="compose_error_no_signing_key">No key configured for signing! Please check your settings.</string>
     <string name="compose_error_private_missing_keys">Private mode is enabled, but some recipients do not have keys!</string>
-    <string name="crypto_mode_disabled">Never Sign or Encrypt.</string>
-    <string name="crypto_mode_sign_only">Always Sign, Never Encrypt.</string>
-    <string name="crypto_mode_opportunistic">Always Sign, Encrypt if possible.</string>
-    <string name="crypto_mode_private">Always Sign, Always Encrypt.</string>
+    <string name="crypto_mode_disabled">Never Encrypt.</string>
+    <string name="crypto_mode_opportunistic">Encrypt if possible.</string>
+    <string name="crypto_mode_private">Always Encrypt.</string>
+    <string name="crypto_mode_supsig_disabled">Never Sign or Encrypt.</string>
+    <string name="crypto_mode_supsig_sign_only">Always Sign, Never Encrypt.</string>
+    <string name="crypto_mode_supsig_opportunistic">Always Sign, Encrypt if possible.</string>
+    <string name="crypto_mode_supsig_private">Always Sign, Always Encrypt.</string>
     <string name="error_crypto_provider_connect">Cannot connect to crypto provider, check your settings or click crypto icon to retry!</string>
     <string name="error_crypto_provider_ui_required">Crypto provider access denied, click crypto icon to retry!</string>
     <string name="error_crypto_inline_attach">PGP/INLINE mode does not support attachments!</string>
